moved the layout component to layouts
