+++ packages/client/src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter } from 'react-router-dom';
-import { SwitchRoutes } from './router/SwitchRoutes';
-import { Layout } from './components/Body/Layout';
+import { SwitchRoutes } from '@/router/SwitchRoutes';
+import Layout from '@/layouts/Layout';
 
 function App() {
   return (
